+++ pages/proyectos.php
@@ -612,6 +612,9 @@ desired effect
                           <th>Tipo</th>
                           <th>Categoría</th>
                           <th>Responsable</th>
+                          <?php if ($rq_sec['admin']=='1') { ?> 
+                            <th>Gerencia</th>
+                          <?php } ?>
                           <th>Prioridad</th>
                           <th>Estado</th>
                           <th>Avance</th>
@@ -622,15 +625,16 @@ desired effect
                         </thead>
                         <tbody>
                             <?php
-                            $query = "SELECT i.*, p.nombre, p.apellido, t.nombre as tipo_nombre FROM proyecto as i 
+                            $query = "SELECT i.*, p.nombre, p.apellido, t.nombre as tipo_nombre, g.nombre as gerencia FROM proyecto as i 
                                       LEFT JOIN persona as p on i.responsable = p.id_persona
+                                      LEFT JOIN gerencia as g on p.gerencia = g.id_gerencia
                                       LEFT JOIN tipo_proyecto as t on i.tipo = t.id
                                       WHERE i.borrado='0' AND i.estado!='4' AND p.borrado ='0' ";
                             // AGREGO EL FILTRO DE GERENCIA DEL USUARIO=CIBERSEGURIDAD O LA GERENCIA DEL REFERENTE
-                            // if ( $per_id_gerencia != 1) {
+                            if ($rq_sec['admin']=='0') {
                               $query = $query . " AND p.gerencia = $per_id_gerencia ";
-                            // }                                        
-                            $sql = mysqli_query($con, $query . " ORDER BY id_proyecto ASC;");
+                            }                                        
+                            $sql = mysqli_query($con, $query . " ORDER BY g.nombre, id_proyecto ASC;");
 
                                 while($row = mysqli_fetch_assoc($sql)){
 
@@ -671,7 +675,9 @@ desired effect
                                     echo '
                                     </td>
                                     <td>'.$row['apellido'].' '.$row['nombre']. '</td>'; 
-
+                                    if ($rq_sec['admin']=='1') {
+                                      echo '<td>' . $row['gerencia'] .'</td>';
+                                    }
 
                                     if($row['prioridad'] == '1'){
                                         echo '<td>Alta</td>';
@@ -814,8 +820,9 @@ desired effect
                                     WHERE i.borrado='0' AND i.estado='4' AND p.borrado = '0' ";
                             // AGREGO EL FILTRO DE GERENCIA DEL USUARIO=CIBERSEGURIDAD O LA GERENCIA DEL REFERENTE
                             // if ( $per_id_gerencia != 1) {
+                            if ($rq_sec['admin']=='0') {
                               $query = $query . " AND p.gerencia = $per_id_gerencia ";
-                            // }                                         
+                            }                                         
 
                           $sql = mysqli_query($con, $query . "ORDER BY id_proyecto ASC");